requirements.txt
pywebview>=4.0.2
pyinstaller>=5.6.2
pillow>=9.0.0 